Started writing spell system, splitted code into multiple files. RIP tests, I'll write them again later
