+++ rpgsaga/saga/src/spells.ts
@@ -1,0 +1,16 @@
+import { Ability } from "./character";
+
+class AttackAbility extends Ability
+{
+    
+}
+
+class BuffAbility extends Ability
+{
+
+}
+
+class DebuffAbility extends Ability
+{
+
+}
