+++ rpgsaga/saga/tests/player.spec.ts
@@ -1,4 +1,5 @@
-import {Character, CharacterClass} from "../src/player"
+import {Character} from "../src/character"
+import { CharacterClass } from "../src/characterClasses";
 
 describe("Testing character class enum", () => {
     it('Testing archer item', () => {
@@ -14,9 +15,6 @@ describe("Testing character class enum", () => {
 
 class MockCharacter extends Character
 {
-    public healthPoints: number;
-    public strength: number = 5;
-
     constructor(name: string, characterClass: CharacterClass, healthPoints: number)
     { 
         super(name, characterClass, healthPoints);
@@ -31,21 +29,5 @@ describe("Testing character class", () => {
         expect(character.class).toEqual(CharacterClass.archer);
         expect(character.healthPoints).toEqual(100);
         expect(character.strength).toEqual(5);
-    });
-    it('Character attack method should work', () => {
-        var mage: Character = new MockCharacter("LifeWeaver", CharacterClass.mage, 100);
-        var archer: Character = new MockCharacter("Hanzo", CharacterClass.archer, 100);
-
-        var message = mage.attack(archer, false);
-        expect(archer.healthPoints).toEqual(95);
-        expect(message).toEqual("LifeWeaver (mage) нанёс 5 урона Hanzo (archer)!");
-    });
-    it('Character attack method should miss', () => {
-        var mage: Character = new MockCharacter("LifeWeaver", CharacterClass.mage, 100);
-        var archer: Character = new MockCharacter("Hanzo", CharacterClass.archer, 100);
-
-        var message = mage.attack(archer, true);
-        expect(archer.healthPoints).toEqual(100);
-        expect(message).toEqual("LifeWeaver (mage) пытался атаковать Hanzo (archer), но промахнулся!");
     });
 });
